First working version
Fix seeks on byte-segregated files.
Fix bit-shift on data in byte-segregated files.
Fix size of values in mueller matrix.
Fix returned point count.

+++ src/drivers/bpf/BpfHeader.hpp
@@ -45,7 +45,7 @@ class ILeStream;
 
 struct BpfMuellerMatrix
 {
-    float m_vals[16];
+    double m_vals[16];
 };
 ILeStream& operator >> (ILeStream& stream, BpfMuellerMatrix& m);
